requirements.txt
streamlit
mysql-connector-python
google-generativeai
python-dotenv
PyPDF2
pillow
pytesseract
SpeechRecognition
python-docx
pyaudio